requirements.txt
CairoSVG==2.5.2
progressbar33==2.4
pyperclip==1.8.2
PyQt5==5.15.7
pytz==2022.1
qrcodegen==1.7.0
requests==2.28.1
